+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-06T12:45:23.925Z",
+    "2026-03-07T11:51:10.719Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {