+++ packages/orq-rc/src/models/operations/fileget.ts
@@ -94,7 +94,7 @@ export const FileGetResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-03-06T12:45:27.421Z",
+    "2026-03-07T11:51:14.248Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {