+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -125,7 +125,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-03-06T12:45:27.421Z",
+    "2026-03-07T11:51:14.248Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {