+++ packages/orq-rc/RELEASES.md
@@ -7499,3 +7499,13 @@ Based on:
 - [typescript v4.5.0-rc.37] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.37 - packages/orq-rc
+
+## 2026-03-07 11:54:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.748.0 (2.852.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.38] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.38 - packages/orq-rc